Dimensional unit systems

Geophysics.bohrreduced

Geophysics.feet

Geophysics.kelvin

Geophysics.kilograms

Geophysics.meters

Geophysics.molecules

Geophysics.moles

Geophysics.rankine

Geophysics.slugs

A UnitSystem is a consistent set of dimensional values selected to accomodate a particular use case or standardization. In total, five fundamental constants kB,ħ,𝘤,μ,mₑ are used to specify a specific unit system. These are the constants of boltzmann, planckreduced, lightspeed, permeability, and electronmass. Different choices of natural units or physical measurements result in a variety of unit systems optimized for many purposes.

Missing docstring.

Missing docstring for Geophysics.UnitSystem. Check Documenter's build log for details.

Metric SI Units

In the Systeme International d'Unites (the SI units) the UnitSystem constants are derived from the most accurate possible physical measurements and a few exactly defined constants. Exact values are the avogadro number, boltzmann constant, planck constant, lightspeed definition, and elementary charge definition.

julia> NA # avogadro
6.02214076e23

julia> kB # boltzmann
1.380649e-23

julia> 𝘩 # planck
6.62607015e-34

julia> 𝘤 # lightspeed
2.99792458e8

julia> 𝘦 # charge
1.602176634e-19

Physical measured values with uncertainty are the electron to proton mass ratio μₑₐ, proton to atomic mass ratio μₚₐ, fine structure constant αinv, the Rydberg R∞ constant, and the Planck mass mP.

julia> μₑₐ
0.0005485799090649074

julia> μₚₐ
1.007276466621

julia> αinv
137.035999084

julia> R∞ # rydberg
1.09737315681601e7

julia> mP # planckmass
2.176434e-8

From these numbers along with the 4π*1e-7 value of the Gaussian unit μ₀, the molar mass constant Mᵤ, planckreduced, permeability, electronmass, and proton to electon mass ratio are computed.

julia> Mᵤ # αinv^2*R∞*NA*2𝘩/𝘤/μₑₐ
0.000999999999656256

julia> ħ # 𝘩/2π
1.0545718176461565e-34

julia> μ₀+δμ₀ # 2𝘩/𝘤/αinv/𝘦^2
1.256637062121048e-6

julia> mₑ # Mᵤ*μₑₐ/NA
9.109383701558256e-31

julia> μₚₑ # μₚₐ/μₑₐ
1836.152673432705

This results in the Metric::UnitSystem{kB,ħ,𝘤,μ₀,mₑ} and SI2019::UnitSystem{kB,ħ,𝘤,μ₀+δμ₀,mₑ} variants of SI.

Missing docstring.

Missing docstring for Geophysics.Metric. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Geophysics.SI2019. Check Documenter's build log for details.

Additional reference values include the ground state hyperfine structure transition frequency of caesium-133 ΔνCs and luminous efficacy Kcd of monochromatic radiation of 540 THz.

julia> ΔνCs
9.19263177e9

julia> Kcd
683.0

Other historic systems

Alternatives to the SI unit system are the centimetre-gram-second CGS::UnitSystem{1e7*kB,1e7*ħ,100𝘤,1e7*μ₀,1000mₑ} and CGS2019::UnitSystem{1e7*kB,1e7*ħ,100𝘤,1e7(μ₀+δμ₀),1000mₑ} variants.

Missing docstring.

Missing docstring for Geophysics.CGS. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Geophysics.CGS2019. Check Documenter's build log for details.

Historically, the josephson and klitzing constants have been used to define Conventional and CODATA derived UnitSystem variants.

julia> josephson(Conventional)
4.835979e14

julia> klitzing(Conventional)
25812.807

julia> josephson(CODATA)
4.835978525e14

julia> klitzing(CODATA)
25812.8074555

Missing docstring.

Missing docstring for Geophysics.Conventional. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Geophysics.CODATA. Check Documenter's build log for details.

In Britain and the United States an English system of engineering units was commonly used.

Missing docstring.

Missing docstring for Geophysics.English. Check Documenter's build log for details.

Natural units

With the introduction of the planckmass a set of natural atomic unit systems can be derived in terms of the gravitational coupling constant.

julia> αG # (mₑ/mP)^2
1.751809945750515e-45

Some of the notable variants include

Planck::UnitSystem{1,1,1,1,√(4π*αG)}
PlanckGauss::UnitSystem{1,1,1,4π,√αG}
Stoney::UnitSystem{1,αinv,1,4π,√(αG*αinv)}
Hartree::UnitSystem{1,1,αinv,4π/αinv^2,1}
Rydberg::UnitSystem{1,1,2αinv,π/αinv^2,1/2}
Schrodinger::UnitSystem{1,1,αinv,4π/αinv^2,√αinv*mₑ/mP}
Electronic::UnitSystem{1,αinv,1,4π,1}
Natural::UnitSystem{1,1,1,1,1}
NaturalGauss::UnitSystem{1,1,1,4π,1}
QCD::UnitSystem{1,1,1,1,1/μₚₑ}
QCDGauss::UnitSystem{1,1,1,4π,1/μₚₑ}
QCDoriginal::UnitSystem{1,1,1,4π/αinv,1/μₚₑ}

Missing docstring.

Missing docstring for Geophysics.Planck. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Geophysics.PlanckGauss. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Geophysics.Stoney. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Geophysics.Hartree. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Geophycics.Rydberg. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Geophysics.Schrodinger. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Geophysics.Electronic. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Geophysics.Natural. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Geophysics.NaturalGauss. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Geophysics.QCD. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Geophyiscs.QCDGauss. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Geophyiscs.QCDoriginal. Check Documenter's build log for details.

Fundamental constants of physics

The following are fundamental constants of physics:

Missing docstring.

Missing docstring for avogadro. Check Documenter's build log for details.

Missing docstring.

Missing docstring for boltzmann. Check Documenter's build log for details.

Missing docstring.

Missing docstring for universal. Check Documenter's build log for details.

Missing docstring.

Missing docstring for planck. Check Documenter's build log for details.

Missing docstring.

Missing docstring for planckreduced. Check Documenter's build log for details.

Missing docstring.

Missing docstring for lightspeed. Check Documenter's build log for details.

Missing docstring.

Missing docstring for planckmass. Check Documenter's build log for details.

Missing docstring.

Missing docstring for atomicmass. Check Documenter's build log for details.

Missing docstring.

Missing docstring for protonmass. Check Documenter's build log for details.

Missing docstring.

Missing docstring for electronmass. Check Documenter's build log for details.

Missing docstring.

Missing docstring for newton. Check Documenter's build log for details.

Missing docstring.

Missing docstring for einstein. Check Documenter's build log for details.

Missing docstring.

Missing docstring for permeability. Check Documenter's build log for details.

Missing docstring.

Missing docstring for permittivity. Check Documenter's build log for details.

Missing docstring.

Missing docstring for coulomb. Check Documenter's build log for details.

Missing docstring.

Missing docstring for stefan. Check Documenter's build log for details.

Missing docstring.

Missing docstring for radiationintensity. Check Documenter's build log for details.

Missing docstring.

Missing docstring for impedance. Check Documenter's build log for details.

Missing docstring.

Missing docstring for charge. Check Documenter's build log for details.

Missing docstring.

Missing docstring for magneton. Check Documenter's build log for details.

Missing docstring.

Missing docstring for conductance. Check Documenter's build log for details.

Missing docstring.

Missing docstring for faraday. Check Documenter's build log for details.

Missing docstring.

Missing docstring for josephson. Check Documenter's build log for details.

Missing docstring.

Missing docstring for magneticflux. Check Documenter's build log for details.

Missing docstring.

Missing docstring for klitzing. Check Documenter's build log for details.

Missing docstring.

Missing docstring for hardtree. Check Documenter's build log for details.

Missing docstring.

Missing docstring for rydberg. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bohr. Check Documenter's build log for details.

Geophysics.bohrreduced — Function

bohrreduced(U::UnitSystem) = electronmass(U)/bohr(U)/1836.152673432705

Reduced Bohr radius including the effect of reduced mass in hydrogen atom (m).

julia> bohrreduced(Metric) # m
2.6253145122281045e-44

Common conversion factors

Common conversion factors include moles, molecules, kilograms, slugs, meters, feet, kelvin, and rankine.

Geophysics.moles — Function

moles(N::Real,U::UnitSystem=Metric) = N/avogadro(U)

Converts the number of molecules N to number of moles (mol).

Geophysics.molecules — Function

molecules(n::Real,U::UnitSystem=Metric) = n*avogadro(U)

Converts the number of moles n to number of molecules (dimensionless).

Geophysics.kilograms — Function

kilograms(m::Real) = 14.593902938825488m

Converts mass m from slugs to kilogram (kg).

Geophysics.slugs — Function

slugs(m::Real) = 0.06852176584918959m

Converts mass m from kilograms to slugs (slug).

Geophysics.meters — Function

meters(d) = 0.3048000001333915d

Converts distance d from feet to meters (m).

Geophysics.feet — Function

feet(d) = 3.2808398935773093d

Converts distance d from meters to feet (ft).

Geophysics.kelvin — Function

kelvin(T) = (5/9)T

Converts temperature T from degrees Rankine to Kelvin (K).

Geophysics.rankine — Function

rankine(T) = (9/5)T

Converts temperature T from Kelvin to degrees Rankine (°R).